Optimized includes and fixed compiler warnings and various other fixes

+++ include/nrpe/packet.hpp
@@ -243,11 +243,11 @@ namespace nrpe {
 			ss << ", payload: " << payload_;
 			return ss.str();
 		}
-		static nrpe::packet create_response(int ret, std::string string, int buffer_length) {
-			return packet(nrpe::data::responsePacket, nrpe::data::version2, ret, string, buffer_length);
+		static nrpe::packet create_response(int ret, std::string string, unsigned int buffer_length) {
+			return packet(nrpe::data::responsePacket, nrpe::data::version2, static_cast<int16_t>(ret), string, buffer_length);
 		}
-		static nrpe::packet create_more_response(int ret, std::string string, int buffer_length) {
-			return packet(nrpe::data::moreResponsePacket, nrpe::data::version2, ret, string, buffer_length);
+		static nrpe::packet create_more_response(int ret, std::string string, unsigned int buffer_length) {
+			return packet(nrpe::data::moreResponsePacket, nrpe::data::version2, static_cast<int16_t>(ret), string, buffer_length);
 		}
 	};
 }
